Cleaned stuff up, copying last commit because its important
Used your method for color mean to test if were on top of a bag, in which case we run the looting. I can make this a method to call from agent. I was thinking we could be running this color mean test in our agent, and if it returns true(when its a function) we can stop movement and run looting and then restart movement. Right now im thinking just lifting all keys and shutting down movement threads, as a trivial solution. But im sure theres a better solution. Also we should only stop movement on a timer like once a second so we dont keep stopping the movement before we can get off the bag.