Theres a lot going on here, normally these commits have very bad structure...this one might be better(yeah right)
AgentTest: Added some small stuff here split up aiming and monitoring health

GetData: A class thats just there for abstraction basically we can call this from mt_logic to get the game data

Getmode: When on a loot bag it believed we were on a portal since the value of portal was set for the mode_mean, i put loot_rect into its own if so it doesnt get stepped on by portals. This works very well.

Movement: This is just your movement code pasted into a new file, i took out the while loops in testing because i dont think we need them, but movement isnt working right now.

mt_logic_Alex: this is just a copy of yours that i added stuff I added more important information to the state and added functionality to run our old agent with new modes. Aiming works well and reading the state works well.

Movement is not working, not sure why yet but the threads stay running and it isnt moving the player, i still have more testing to do on it but this is what i got done so far.