Kind of a lot here...not really though.
Movement is fixed, and working with the agent. Looting is working with the agent and potions are also added. Agent works pretty good, will work excellent once we have nexus features added.

Also made a big breakthrough in terms of object detection for pathfinding. Ive found the colors for trees and dark water on the map, im hoping we can use this information to make a path using the map, if this works i can also get the colors for all the walls and things using the maps you downloaded before.

I believe that is everything i did.