Looting.py: Added a move to the end of looting to move it back into the realm mode.
mt_logic_Alex.py:
Updated movement calls directly to movement class.

Nexus.py:
Added keyUps to the end in case any of them exit mid-press. I think one got caught when it entered a portal without moving vertical at all.

Movement.py:
Created movement class that has the gamestate object in it to read mode directly. Also added a listener for mode changes and consolidated the movement function to make it more generic. Had to add a .1 second sleep because it was slowing down aiming and the game state reads. I think we might be reaching the limit of our single thread. Might have to look at multiprocessing library with shared memory. We can share an array of enemy locations or something so I think this is possible. Might look into it.