BulletTrackingTest.py: Split up functions and added lists of the bullet colors for T0 and T1 weapons (will add more later). I have it ID our bullets based on area > 220 (lowest typically around 230) and matching the contour centers. Seems to be working well. Now it removes our bullets from the contours.
FrameSequenceOut.py:
In Archive. Wrote a script to dump 20 frames continuously with .1sec difference to see how bullets progress across frames. I'm thinking of checking 2 zones around the character rather than trying to track bullets with movement. Look at the number of bullets in zone 1 (outer) and compare next frame with zone 2 (inner) to see if the bullets are moving closer (I can try to explain better when we talk). I think trying to track the trajectory of all bullets is going to be too complicated and unnecessary, the zones solution is a bit simpler. The function that gets the center of the contours can be used to determine if the bullets are in the zones.

Resources Folder:
Added pictures for T0 and T1 staffs (staves?). Also added a folder called FrameSeq in the TestImages folder for the output of the 20 sequential frames from FrameSequenceOut.py