Not a lot changed.
Moved monitoring health to its own thread in an attempt to speed it up so we dont die as often. seems to work well.

Also experimenting with template matching bags off the frame(probably wont do this) I let it run for about 30 minutes and it missed a lot of bags and i think it wouldve leveled up a lot faster with better gear. so i was thinking of finding a way to move to bags on screen if there arent enemies around.