Messed with data values for retreating and moving towards in relation to movement.
Made the agent much more aggresive in its pursuit to loot.

Also added a method to test if were out of potions and if we are we jump to the nexus at a higher health than if we had potions didnt have time to test this yet. Will follow up with testing tonight.